+++ CHANGELOG.rst
@@ -20,6 +20,8 @@ New features and notable changes:
 
 - Add support for Markdown output format. (:issue:`1072`)
 
+  - Add option :option:`--markdown-file-link` to link files in markdown report. (:issue:`1079`)
+
 Bug fixes and small improvements:
 
 - Fix warning ``Deprecated config key None used, please use 'txt-metric=branch' instead.``